requirements.txt
# Core dependencies
pandas>=1.5.0
numpy>=1.23.0

# Web scraping
requests>=2.28.0
beautifulsoup4>=4.11.0
selenium>=4.8.0
webdriver-manager>=3.8.0

# LinkedIn scraping
linkedin-scraper>=0.1.0

# PDF processing
pdfplumber>=0.9.0

# Environment variables
python-dotenv>=1.0.0

# Email (built-in, but listed for clarity)
# smtplib - built-in
# email - built-in

# Optional: For better table display (if tabulate is available)
tabulate>=0.9.0

